requirements.txt
obs-websocket-py==1.0
watchdog==3.0.0
Pillow>=10.3.0
requests>=2.31.0
psutil>=5.9.0
Flask>=3.0.0
Flask-SocketIO>=5.3.0
python-socketio>=5.10.0
